requirements.txt
# requirements.txt
requests>=2.26.0
pandas>=1.3.0
python-dotenv>=0.19.0
tqdm>=4.62.0
transformers>=4.26.0
torch>=2.0.0